backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
Pillow==10.2.0
psycopg2-binary==2.9.9
python-decouple==3.8
django-cloudinary-storage==0.3.0
cloudinary==1.36.0
django-taggit==4.0.0
django-ckeditor==6.7.0
gunicorn==21.2.0
whitenoise==6.6.0 